[SERVICE-478] Seperate AppWindow implementation and interface
